move elements in modules + fix anchors
