+++ src/app/core/footer/footer.component.html
@@ -1,4 +1,5 @@
-<footer>
+<a id="contact" class="anchor"></a>
+<footer id="contact-summary">
   <div class="light-blue-darken-1 py-3">
     <div class="container">
       <div class="row">
