+++ src/app/core/header/header.component.html
@@ -3,10 +3,10 @@
     <div class="logo ml-4 mr-2"></div>
     <div>JOSCELYN JEAN</div>
     <div class="ml-auto mr-4">
-      <button mat-button>Community</button>
-      <button mat-button>Projects</button>
-      <button mat-button>Contact</button>
-      <button mat-raised-button color="primary" class="ml-3">Français</button>
+      <a mat-button routerLink="summary" fragment="community">Community</a>
+      <a mat-button routerLink="summary" fragment="projects">Projects</a>
+      <a mat-button routerLink="summary" fragment="contact">Contact</a>
+      <a mat-raised-button color="primary" class="ml-3">Français</a>
     </div>
   </div>
 </header>
