+++ src/app/modules/community/community.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { TwitterComponent } from './twitter/twitter.component';
 import { StackoverflowComponent } from './stackoverflow/stackoverflow.component';
 import { CommunitySummaryComponent } from './shared/community-summary/community-summary.component';
-import { AppMaterialModule } from '../app-material.module';
+import { AppMaterialModule } from '../../app-material.module';
 
 @NgModule({
   declarations: [TwitterComponent, StackoverflowComponent, CommunitySummaryComponent],
