+++ src/app/modules/community/shared/community-summary/community-summary.component.html
@@ -1,3 +1,4 @@
+<a id="community" class="anchor"></a>
 <div id="community-summary" class="border-top border-bottom border-dark text-left py-4">
   <div class="container">
     <h1><i class="far fa-comments"></i> In the community</h1>
