+++ src/app/modules/community/shared/community-summary/community-summary.component.scss
