+++ src/app/modules/profile/shared/profile-summary/profile-summary.component.scss
@@ -11,7 +11,7 @@ img {
 
 #profile-summary {
   padding-top: 58px;
-  background-image: url('../../../../assets/images/banner.jpg');
+  background-image: url('../../../../../assets/images/banner.jpg');
   background-repeat: no-repeat;
   background-size: cover;
   background-position-x: center;
