+++ src/app/modules/projects/project.module.ts
@@ -1,4 +1,4 @@
-import { AppMaterialModule } from './../app-material.module';
+import { AppMaterialModule } from './../../app-material.module';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ProjectSummaryComponent } from './shared/project-summary/project-summary.component';
