+++ src/app/modules/projects/shared/project-summary/project-summary.component.html
@@ -1,3 +1,4 @@
+<a id="projects" class="anchor"></a>
 <div id="project-summary" class="text-left py-4">
   <div class="container">
     <h1 class="mb-4"><i class="fab fa-github"></i> My Github projects</h1>
