+++ src/app/modules/projects/shared/project-summary/project-summary.component.scss
